#22 - Made changes according to feedback
